refactor: reorganize service and client classes
- Move WAF-specific logic from `CloudflareService` to new `WafRuleService`
- Make `CloudflareService` abstract base class for feature-specific services
- Move WAF API methods from `CloudflareApiClient` to new `WafApiClient`
- Make `CloudflareApiClient` abstract base class for feature-specific client calls
- Update environment variables from `CF_WAF_*` to `CFCACHE_*`
- Rename config from `cf-waf-rule.php` to `cfcache.php`
