+++ src/Http/CloudflareApiClient.php
@@ -205,93 +205,6 @@ class CloudflareApiClient
     }
 
     /**
-     * Fetch all firewall rules for the configured zone.
-     *
-     * @return array Array of firewall rule data
-     *
-     * @throws CloudflareApiException If the request fails
-     */
-    public function getFirewallRules(): array
-    {
-        $response = $this->get("zones/{$this->zoneId}/firewall/rules");
-
-        return $response->json('result', []);
-    }
-
-    /**
-     * Fetch a specific firewall rule by ID.
-     *
-     * @param  string  $ruleId  The firewall rule ID
-     * @return array The firewall rule data
-     *
-     * @throws CloudflareApiException If the request fails
-     */
-    public function getFirewallRule(string $ruleId): array
-    {
-        $response = $this->get("zones/{$this->zoneId}/firewall/rules/{$ruleId}");
-
-        return $response->json('result', []);
-    }
-
-    /**
-     * Create one or more firewall rules.
-     *
-     * @param  array  $rules  Array of rule definitions
-     * @return array The created firewall rules
-     *
-     * @throws CloudflareApiException If the request fails
-     */
-    public function createFirewallRules(array $rules): array
-    {
-        $response = $this->post("zones/{$this->zoneId}/firewall/rules", $rules);
-
-        return $response->json('result', []);
-    }
-
-    /**
-     * Create one or more filters.
-     *
-     * @param  array  $filters  Array of filter definitions
-     * @return array The created filters
-     *
-     * @throws CloudflareApiException If the request fails
-     */
-    public function createFilters(array $filters): array
-    {
-        $response = $this->post("zones/{$this->zoneId}/filters", $filters);
-
-        return $response->json('result', []);
-    }
-
-    /**
-     * Update a filter.
-     *
-     * @param  string  $filterId  The filter ID to update
-     * @param  array  $filterData  The updated filter data
-     * @return array The updated filter
-     *
-     * @throws CloudflareApiException If the request fails
-     */
-    public function updateFilter(string $filterId, array $filterData): array
-    {
-        $response = $this->put("zones/{$this->zoneId}/filters/{$filterId}", $filterData);
-
-        return $response->json('result', []);
-    }
-
-    /**
-     * Delete a filter.
-     *
-     * @param  string  $filterId  The filter ID to delete
-     *
-     * @throws CloudflareApiException If the request fails
-     */
-    public function deleteFilter(string $filterId): void
-    {
-        $this->delete("zones/{$this->zoneId}/filters/{$filterId}");
-    }
-
-    /**
      * Get the configured zone ID.
      *
      * @return string The Cloudflare zone ID
