+++ tests/Feature/Http/CloudflareApiClientTest.php
@@ -186,117 +186,6 @@ class CloudflareApiClientTest extends TestCase
     }
 
     #[Test]
-    public function it_returns_parsed_result(): void
-    {
-        $expectedRules = [
-            ['id' => 'rule-1', 'description' => 'Rule 1'],
-            ['id' => 'rule-2', 'description' => 'Rule 2'],
-        ];
-
-        Http::fake([
-            '*' => Http::response(['success' => true, 'result' => $expectedRules], 200),
-        ]);
-
-        $rules = $this->client->getFirewallRules();
-
-        $this->assertEquals($expectedRules, $rules);
-    }
-
-    #[Test]
-    public function it_returns_single_rule(): void
-    {
-        $expectedRule = ['id' => 'rule-123', 'description' => 'Test Rule'];
-
-        Http::fake([
-            '*' => Http::response(['success' => true, 'result' => $expectedRule], 200),
-        ]);
-
-        $rule = $this->client->getFirewallRule('rule-123');
-
-        $this->assertEquals($expectedRule, $rule);
-
-        Http::assertSent(function (Request $request) {
-            return str_contains($request->url(), 'firewall/rules/rule-123');
-        });
-    }
-
-    #[Test]
-    public function it_returns_created_rules(): void
-    {
-        $rulesToCreate = [
-            ['action' => 'block', 'description' => 'Rule 1'],
-        ];
-        $expectedResult = [
-            ['id' => 'new-rule-1', 'action' => 'block', 'description' => 'Rule 1'],
-        ];
-
-        Http::fake([
-            '*' => Http::response(['success' => true, 'result' => $expectedResult], 200),
-        ]);
-
-        $result = $this->client->createFirewallRules($rulesToCreate);
-
-        $this->assertEquals($expectedResult, $result);
-    }
-
-    #[Test]
-    public function it_returns_created_filters(): void
-    {
-        $filtersToCreate = [
-            ['expression' => 'test expression', 'description' => 'Test filter'],
-        ];
-        $expectedResult = [
-            ['id' => 'filter-1', 'expression' => 'test expression', 'description' => 'Test filter'],
-        ];
-
-        Http::fake([
-            '*' => Http::response(['success' => true, 'result' => $expectedResult], 200),
-        ]);
-
-        $result = $this->client->createFilters($filtersToCreate);
-
-        $this->assertEquals($expectedResult, $result);
-    }
-
-    #[Test]
-    public function it_returns_updated_filter(): void
-    {
-        $filterData = [
-            'id' => 'filter-123',
-            'expression' => 'updated expression',
-            'description' => 'Updated filter',
-        ];
-
-        Http::fake([
-            '*' => Http::response(['success' => true, 'result' => $filterData], 200),
-        ]);
-
-        $result = $this->client->updateFilter('filter-123', $filterData);
-
-        $this->assertEquals($filterData, $result);
-
-        Http::assertSent(function (Request $request) {
-            return str_contains($request->url(), 'filters/filter-123')
-                && $request->method() === 'PUT';
-        });
-    }
-
-    #[Test]
-    public function it_sends_delete_request(): void
-    {
-        Http::fake([
-            '*' => Http::response([], 204),
-        ]);
-
-        $this->client->deleteFilter('filter-123');
-
-        Http::assertSent(function (Request $request) {
-            return str_contains($request->url(), 'filters/filter-123')
-                && $request->method() === 'DELETE';
-        });
-    }
-
-    #[Test]
     public function it_includes_endpoint_and_method_information(): void
     {
         Http::fake([
