+++ workbench/.env.example
@@ -8,5 +8,5 @@ APP_DEBUG=true
 # `vendor/bin/testbench key:generate --show`
 APP_KEY=
 
-CF_WAF_API_TOKEN=
-CF_WAF_ZONE_ID=
+CFCACHE_API_TOKEN=
+CFCACHE_ZONE_ID=
